moodshare/requirements.txt
# =============================================================================
# MOODSHARE DEPENDENCIES
# =============================================================================
# Updated for Python 3.13 compatibility

# Core Framework
Flask>=2.3.0
Werkzeug>=3.0.0

# Database ORM (updated for Python 3.13)
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.23

# Database Migrations
Flask-Migrate>=4.0.5
alembic>=1.13.0

# Authentication
Flask-Login>=0.6.3

# Forms & Validation
Flask-WTF>=1.2.1
WTForms>=3.1.0
email-validator>=2.1.0

# Email Support
Flask-Mail>=0.9.1

# Date/Time Formatting
Flask-Moment>=1.0.5

# JWT for password reset tokens
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Environment Variables
python-dotenv>=1.0.0
